+++ src/Schema/Events/Toggle.ts
@@ -37,7 +37,7 @@ export interface ToggledNotification {
 }
 
 /**
- * A user toggles a saved search on/off on iOS
+ * A user toggles a saved search on/off on an app or web
  *
  *  This schema describes events sent to Segment from [[toggledNotification]]
  *
@@ -59,7 +59,7 @@ export interface ToggledSavedSearch {
   context_screen_owner_type: ScreenOwnerType
   context_screen_owner_id?: string
   context_screen_owner_slug?: string
-  modified: boolean
-  original: boolean
+  modified?: boolean
+  original?: boolean
   search_criteria_id: string
 }