+++ src/Schema/Values/ContextModule.ts
@@ -34,6 +34,7 @@ export enum ContextModule {
   auctionResults = "auctionResults",
   auctionResultsForArtistsYouFollow = "auctionResultsForArtistsYouFollow",
   auctionResultsRail = "auctionResultsRail",
+  auctionResultComparableWorks = "auctionResultComparableWorks",
   auctionSidebar = "auctionSidebar",
   auctionsInfo = "auctionsInfo",
   auctionTab = "auctionTab",
@@ -103,6 +104,7 @@ export enum ContextModule {
   ordersRespond = "ordersRespond",
   ordersReview = "ordersReview",
   ordersShipping = "ordersShipping",
+  ordersSubmitted = "ordersSubmitted",
   otherCollectionsRail = "otherCollectionsRail",
   otherWorksByArtistRail = "otherWorksByArtistRail",
   otherWorksFromPartnerRail = "otherWorksFromPartnerRail",
@@ -175,6 +177,7 @@ export type AuthContextModule =
   | ContextModule.auctionRail
   | ContextModule.auctionResult
   | ContextModule.auctionResults
+  | ContextModule.auctionResultComparableWorks
   | ContextModule.auctionSidebar
   | ContextModule.auctionsInfo
   | ContextModule.auctionTab